Move InspectionPersonsNumberCard to components/index/cards
