+++ components/index/cards/InspectionPersonsNumberCard.vue
